feat(transaction): add PayTag transfers and reconciliation

+++ transaction-service/src/main/java/com/paypal/transaction_service/dto/dto/DebitRequest.java
@@ -1,9 +1,11 @@
 package com.paypal.transaction_service.dto.dto;
 
+import java.math.BigDecimal;
+
 public class DebitRequest {
     private Long userId;
     private String currency;
-    private Long amount;
+    private BigDecimal amount;
 
     public Long getUserId() { return userId; }
     public void setUserId(Long userId) { this.userId = userId; }
@@ -11,6 +13,6 @@ public class DebitRequest {
     public String getCurrency() { return currency; }
     public void setCurrency(String currency) { this.currency = currency; }
 
-    public Long getAmount() { return amount; }
-    public void setAmount(Long amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
 }
