+++ transaction-service/src/main/java/com/paypal/transaction_service/dto/dto/HoldRequest.java
@@ -1,11 +1,12 @@
 package com.paypal.transaction_service.dto.dto;
 
 
+import java.math.BigDecimal;
 
 public class HoldRequest {
     private Long userId;
     private String currency;
-    private Long amount;
+    private BigDecimal amount;
 
     public Long getUserId() { return userId; }
     public void setUserId(Long userId) { this.userId = userId; }
@@ -13,6 +14,6 @@ public class HoldRequest {
     public String getCurrency() { return currency; }
     public void setCurrency(String currency) { this.currency = currency; }
 
-    public Long getAmount() { return amount; }
-    public void setAmount(Long amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
 }
